// src/app/externalApis/models/BotInfoMediasInner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IM
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BotInfoMediasInner
 */
export interface BotInfoMediasInner {
    /**
     * 
     * @type {string}
     * @memberof BotInfoMediasInner
     */
    link: string;
    /**
     * 
     * @type {string}
     * @memberof BotInfoMediasInner
     */
    media: BotInfoMediasInnerMediaEnum;
}


/**
 * @export
 */
export const BotInfoMediasInnerMediaEnum = {
    OfficialWebsite: 'Official website',
    X: 'X',
    Ontello: 'Ontello',
    Discord: 'Discord'
} as const;
export type BotInfoMediasInnerMediaEnum = typeof BotInfoMediasInnerMediaEnum[keyof typeof BotInfoMediasInnerMediaEnum];


/**
 * Check if a given object implements the BotInfoMediasInner interface.
 */
export function instanceOfBotInfoMediasInner(value: object): value is BotInfoMediasInner {
    if (!('link' in value) || value['link'] === undefined) return false;
    if (!('media' in value) || value['media'] === undefined) return false;
    return true;
}

export function BotInfoMediasInnerFromJSON(json: any): BotInfoMediasInner {
    return BotInfoMediasInnerFromJSONTyped(json, false);
}

export function BotInfoMediasInnerFromJSONTyped(json: any, ignoreDiscriminator: boolean): BotInfoMediasInner {
    if (json == null) {
        return json;
    }
    return {
        
        'link': json['link'],
        'media': json['media'],
    };
}

export function BotInfoMediasInnerToJSON(json: any): BotInfoMediasInner {
    return BotInfoMediasInnerToJSONTyped(json, false);
}

export function BotInfoMediasInnerToJSONTyped(value?: BotInfoMediasInner | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'link': value['link'],
        'media': value['media'],
    };
}